Decompose metagame_analysis handlers into focused mixins
Split the 349-LOC MetagameAnalysisHandlersMixin into three low-coupling
mixins matching the clusters identified in the architecture review:

- navigation.py (MetagameNavigationMixin): toolbar callbacks, control
  sync, busy-state, on_close
- data_loader.py (MetagameDataLoaderMixin): worker-thread fetch, error
  handling, stats population, default-offset selection
- visualization.py (MetagameVisualizationMixin): pie-chart drawing and
  changes-panel rendering

handlers.py now composes the three mixins under the existing
MetagameAnalysisHandlersMixin name, so frame.py and external imports are
unchanged. No behavior changes; method bodies were moved verbatim.
